fixed fp errors: changed floatrs from 32 to 64 bit

+++ quorc/src/frontend/ast.rs
@@ -119,7 +119,7 @@ pub enum Expr {
     SizeOf(SizeOf),
     IntLiteral(i32),
     LongLiteral(i64),
-    FloatLiteral(f32),
+    FloatLiteral(f64),
     BoolLiteral(bool),
     StringLiteral(String),
     // name of class
